backend/anime_generator.py
import tensorflow as tf
from PIL import Image
import numpy as np
import os

# Load a pre-trained model (placeholder for actual model path)
# In practice, you would replace this with the path to your actual model file or use a model loaded from TensorFlow Hub
model_path = 'path_to_your_pretrained_model.h5'
model = tf.keras.models.load_model(model_path, compile=False)

def prepare_image(image_path):
    """ Load an image file and prepare it for the model. """
    img = Image.open(image_path)
    img = img.resize((256, 256))  # Resize to the input size expected by the model
    img = np.array(img)
    img = (img / 127.5) - 1  # Normalize the image to [-1, 1]
    img = np.expand_dims(img, axis=0)  # Add batch dimension
    return img

def generate_anime_style(frames):
    """ Convert frames to anime style using a pre-trained model. """
    anime_frames = []
    for frame_path in frames:
        img = prepare_image(frame_path)
        anime_img = model.predict(img)  # Run the model prediction
        anime_img = (anime_img + 1) * 127.5  # Convert back to [0,255]
        anime_img = anime_img.astype(np.uint8)
        anime_img = np.squeeze(anime_img)  # Remove batch dimension
        anime_img_pil = Image.fromarray(anime_img)
        anime_output_path = f"{os.path.splitext(frame_path)[0]}_anime.png"
        anime_img_pil.save(anime_output_path)  # Save the anime style frame
        anime_frames.append(anime_output_path)
    return anime_frames

# Example usage:
# frames = ["path_to_frame1.jpg", "path_to_frame2.jpg"]
# anime_frames = generate_anime_style(frames)
# print("Generated anime style frames:", anime_frames)
