requirements.txt
Flask==2.0.2
numpy==1.21.2
Pillow==8.3.2
opencv-python==4.5.3.56
tensorflow==2.6.0
gunicorn==20.1.0
